refactor(monitor): [#13] FE 경계와 그래프 뷰 분해를 정리한다
