+++ src/entities/run/index.ts
@@ -1,2 +1,7 @@
 export * from "./lib/selectors";
+export {
+  FIXTURE_DATASETS,
+  FIXTURE_IMPORT_TEXT,
+  LIVE_FIXTURE_FRAMES,
+} from "./model/samples";
 export * from "./model/types";
