--- src/entities/run/testing.ts
@@ -1,5 +1,0 @@
-export {
-  FIXTURE_DATASETS,
-  FIXTURE_IMPORT_TEXT,
-  LIVE_FIXTURE_FRAMES,
-} from "./model/samples";
