+++ src/pages/monitor/index.ts
@@ -1,16 +1,1 @@
-export { isEditableKeyboardTarget } from "./lib/keyboardTarget";
-export { dispatchMonitorKeyboardShortcut } from "./lib/useMonitorKeyboardShortcuts";
-export { createMonitorArchiveActions } from "./model/createMonitorArchiveActions";
-export { createMonitorImportExportActions } from "./model/createMonitorImportExportActions";
-export { createMonitorViewActions } from "./model/createMonitorViewActions";
-export { deriveMonitorViewState } from "./model/deriveMonitorViewState";
-export type {
-  MonitorAction,
-  MonitorState,
-  SelectionLoadState,
-} from "./model/state";
-export {
-  createMonitorInitialState,
-  monitorStateReducer,
-} from "./model/state";
 export { MonitorPage } from "./ui/MonitorPage";
