+++ src/pages/monitor/model/state/datasetState.ts
@@ -1,4 +1,4 @@
-import { LIVE_FIXTURE_FRAMES } from "../../../../entities/run/testing";
+import { LIVE_FIXTURE_FRAMES } from "../../../../entities/run";
 import { applyLiveFrame } from "../../../../features/follow-live";
 import {
   buildCollapsedGapIds,
