+++ src/pages/monitor/model/useMonitorPageState.ts
@@ -1,26 +1,16 @@
 import {
-  startTransition,
-  useEffect,
-  useEffectEvent,
   useReducer,
-  useRef,
 } from "react";
-import { LIVE_FIXTURE_FRAMES } from "../../../entities/run/testing";
-import {
-  loadArchivedSessionIndex,
-  loadRecentSessionIndex,
-  loadRecentSessionSnapshot,
-} from "../../../entities/session-log";
-import { canInvokeTauriRuntime } from "../../../shared/api";
 import { useMonitorKeyboardShortcuts } from "../lib/useMonitorKeyboardShortcuts";
 import { createMonitorActions } from "./createMonitorActions";
 import { deriveMonitorViewState } from "./deriveMonitorViewState";
 import {
-  ARCHIVE_PAGE_SIZE,
   createMonitorInitialState,
-  LIVE_FIXTURE_TRACE_ID,
   monitorStateReducer,
 } from "./state";
+import { useLiveFixturePlayback } from "./useLiveFixturePlayback";
+import { useMonitorBootstrap } from "./useMonitorBootstrap";
+import { useMonitorRequestController } from "./useMonitorRequestController";
 
 export function useMonitorPageState() {
   const [state, dispatch] = useReducer(
@@ -28,153 +18,33 @@ export function useMonitorPageState() {
     undefined,
     createMonitorInitialState,
   );
-  const recentSnapshotRequestIdRef = useRef(0);
-  const archiveIndexRequestIdRef = useRef(0);
-  const archiveSnapshotRequestIdRef = useRef(0);
   const derivedState = deriveMonitorViewState(state);
-
-  const cancelPendingSelectionLoad = useEffectEvent(() => {
-    const nextRecentRequestId = recentSnapshotRequestIdRef.current + 1;
-    recentSnapshotRequestIdRef.current = nextRecentRequestId;
-    dispatch({
-      type: "cancel-recent-snapshot-request",
-      requestId: nextRecentRequestId,
-    });
-
-    const nextArchivedRequestId = archiveSnapshotRequestIdRef.current + 1;
-    archiveSnapshotRequestIdRef.current = nextArchivedRequestId;
-    dispatch({
-      type: "cancel-archived-snapshot-request",
-      requestId: nextArchivedRequestId,
-    });
+  const {
+    loadArchiveIndex,
+    requestArchiveIndex,
+    requestRecentIndex,
+    requestRecentSnapshot,
+    searchArchive,
+    selectArchivedSession,
+  } = useMonitorRequestController({
+    state,
+    dispatch,
   });
 
-  const requestRecentIndex = useEffectEvent(() => {
-    dispatch({ type: "begin-recent-index-request" });
-
-    loadRecentSessionIndex().then((items) => {
-      if (items === null) {
-        dispatch({
-          type: "finish-recent-index-request",
-          error: "Recent sessions are unavailable right now.",
-        });
-        return;
-      }
-
-      startTransition(() => {
-        dispatch({ type: "resolve-recent-index-request", items });
-      });
-    });
+  useMonitorBootstrap({
+    activeDataset: derivedState.activeDataset,
+    recentIndex: state.recentIndex,
+    recentIndexReady: state.recentIndexReady,
+    recentSnapshotLoadingId: state.recentSnapshotLoadingId,
+    requestArchiveIndex,
+    requestRecentIndex,
+    requestRecentSnapshot,
   });
-
-  const requestRecentSnapshot = useEffectEvent((filePath: string) => {
-    cancelPendingSelectionLoad();
-
-    const cachedDataset = state.hydratedDatasetsByFilePath[filePath];
-    if (cachedDataset) {
-      dispatch({ type: "set-active-run", traceId: cachedDataset.run.traceId });
-      return;
-    }
-
-    const requestId = recentSnapshotRequestIdRef.current + 1;
-    recentSnapshotRequestIdRef.current = requestId;
-    dispatch({ type: "begin-recent-snapshot-request", requestId, filePath });
-
-    loadRecentSessionSnapshot(filePath).then((dataset) => {
-      if (!dataset) {
-        dispatch({ type: "finish-recent-snapshot-request", requestId });
-        return;
-      }
-
-      dispatch({
-        type: "begin-recent-snapshot-build",
-        requestId,
-        filePath,
-      });
-      startTransition(() => {
-        dispatch({
-          type: "resolve-recent-snapshot-request",
-          requestId,
-          filePath,
-          dataset,
-        });
-      });
-    });
+  useLiveFixturePlayback({
+    datasets: state.datasets,
+    appliedLiveFrames: state.appliedLiveFrames,
+    dispatch,
   });
-
-  const requestArchiveIndex = useEffectEvent(
-    (offset: number, append: boolean, search?: string) => {
-      const requestId = archiveIndexRequestIdRef.current + 1;
-      archiveIndexRequestIdRef.current = requestId;
-      dispatch({ type: "begin-archived-index-request", requestId });
-
-      loadArchivedSessionIndex(offset, ARCHIVE_PAGE_SIZE, search).then((result) => {
-        if (!result) {
-          dispatch({
-            type: "finish-archived-index-request",
-            requestId,
-            error: "Archive sessions are unavailable right now.",
-          });
-          return;
-        }
-
-        startTransition(() => {
-          dispatch({
-            type: "resolve-archived-index-request",
-            requestId,
-            result,
-            append,
-          });
-        });
-      });
-    },
-  );
-
-  useEffect(() => {
-    if (!canInvokeTauriRuntime()) {
-      return;
-    }
-
-    requestRecentIndex();
-  }, []);
-
-  useEffect(() => {
-    if (
-      !state.recentIndexReady ||
-      !state.recentIndex.length ||
-      state.recentSnapshotLoadingId ||
-      derivedState.activeDataset
-    ) {
-      return;
-    }
-
-    requestRecentSnapshot(state.recentIndex[0].filePath);
-  }, [
-    state.recentIndexReady,
-    state.recentIndex,
-    state.recentSnapshotLoadingId,
-    derivedState.activeDataset,
-  ]);
-
-  useEffect(() => {
-    const liveFixtureRun = state.datasets.find(
-      (item) => item.run.traceId === LIVE_FIXTURE_TRACE_ID,
-    );
-    if (!liveFixtureRun || liveFixtureRun.run.liveMode !== "live") {
-      return undefined;
-    }
-
-    if (state.appliedLiveFrames >= LIVE_FIXTURE_FRAMES.length) {
-      return undefined;
-    }
-
-    const frame = LIVE_FIXTURE_FRAMES[state.appliedLiveFrames];
-    const timeout = window.setTimeout(() => {
-      dispatch({ type: "apply-live-frame" });
-    }, frame.delayMs);
-
-    return () => window.clearTimeout(timeout);
-  }, [state.appliedLiveFrames, state.datasets]);
 
   useMonitorKeyboardShortcuts({
     dispatch,
@@ -182,14 +52,6 @@ export function useMonitorPageState() {
     selection: state.selection,
     graphRows: derivedState.graphScene.rows,
   });
-
-  useEffect(() => {
-    if (!canInvokeTauriRuntime()) {
-      return;
-    }
-
-    requestArchiveIndex(0, false);
-  }, []);
 
   return {
     state,
@@ -199,10 +61,10 @@ export function useMonitorPageState() {
       dispatch,
       activeDataset: derivedState.activeDataset,
       activeFollowLive: derivedState.activeFollowLive,
+      loadArchiveIndex,
+      searchArchive,
+      selectArchivedSession,
       requestRecentSnapshot,
-      requestArchiveIndex,
-      archiveSnapshotRequestIdRef,
-      cancelPendingSelectionLoad,
     }),
   };
 }
